requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.15.0  # via google-cloud-core, google-cloud-datastore
google-auth==1.10.0       # via google-api-core
google-cloud-core==1.1.0  # via google-cloud-datastore
google-cloud-datastore==1.10.0  # via google-cloud-ndb
google-cloud-ndb==0.2.1
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.26.0            # via google-api-core
idna==2.8                 # via requests
protobuf==3.11.2          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.7     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-dotenv==0.10.3
pytz==2019.3              # via google-api-core
redis==3.3.11             # via google-cloud-ndb
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.13.0               # via google-api-core, google-auth, grpcio, protobuf
urllib3==1.25.7           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
